Split apart files, keeping autogen-shaped data representations in main/autogen/data
