+++ main/autogen/BUILD
@@ -3,13 +3,11 @@ cc_library(
     srcs = [
         "autogen.cc",
         "autoloader.cc",
-        "msgpack.cc",
         "subclasses.cc",
     ],
     hdrs = [
         "autogen.h",
         "autoloader.h",
-        "msgpack.h",
         "subclasses.h",
     ],
     linkstatic = select({
@@ -21,8 +19,8 @@ cc_library(
         "//ast",
         "//ast/treemap",
         "//core",
+        "//main/autogen/data",
         "//main/options",
         "@com_github_d_bahr_crcpp",
-        "@com_github_ludocode_mpack",
     ],
 )
