+++ main/autogen/data/msgpack.h
@@ -3,7 +3,7 @@
 // has to go first because it violates our poisons
 #include "mpack/mpack.h"
 
-#include "main/autogen/autogen.h"
+#include "main/autogen/data/definitions.h"
 
 namespace sorbet::autogen {
